Consolidate different articles under `layout`
- Brings AR conventions in line with DCR
- Moves the disparate article layouts under the `layout` component
- Moves `labsLogo` and `cutout` to the top level of components
