+++ apps-rendering/src/components/cutout.tsx
